# Conflicts:
#	remote-api/src/test/java/org/alfresco/AppContext04TestSuite.java

+++ remote-api/src/main/java/org/alfresco/rest/api/impl/NodesImpl.java
@@ -2401,6 +2401,8 @@ public class NodesImpl implements Nodes
         NodePermissions nodePerms = nodeInfo.getPermissions();
         if (nodePerms != null)
         {
+            String siteManagerAuthority = getSiteManagerAuthority(nodeRef);
+
             // Cannot set inherited permissions, only direct (locally set) permissions can be set
             if ((nodePerms.getInherited() != null) && (nodePerms.getInherited().size() > 0))
             {
@@ -2411,7 +2413,7 @@ public class NodesImpl implements Nodes
             if (nodePerms.getIsInheritanceEnabled() != null)
             {
                 // If inheritance flag is being disabled, the site manager needs to have permission
-                setSiteManagerPermission(nodeRef, nodePerms);
+                setSiteManagerPermission(nodeRef, nodePerms, siteManagerAuthority);
 
                 if (nodePerms.getIsInheritanceEnabled() != permissionService.getInheritParentPermissions(nodeRef))
                 {
@@ -2510,8 +2512,19 @@ public class NodesImpl implements Nodes
                 }
 
                 // remove any remaining direct perms
+                boolean isInheritanceEnabled = permissionService.getInheritParentPermissions(nodeRef);
                 for (AccessPermission accessPerm : directPerms)
                 {
+                    // prevents deletion of site manager permissions when inheritance is disabled
+                    boolean isSiteManagerAuthority = siteManagerAuthority != null && siteManagerAuthority.equals(accessPerm.getAuthority());
+                    if (!isInheritanceEnabled && isSiteManagerAuthority)
+                    {
+                        if (logger.isDebugEnabled())
+                        {
+                            logger.debug("Site manager permissions will be kept since inheritance flag is disabled: " + nodeRef.getId());
+                        }
+                        continue;
+                    }
                     permissionService.deletePermission(nodeRef, accessPerm.getAuthority(), accessPerm.getPermission());
                 }
             }
@@ -2776,25 +2789,29 @@ public class NodesImpl implements Nodes
         return updateExistingFile(null, nodeRef, fileName, contentInfo, stream, parameters, versionMajor, versionComment);
     }
 
-    private void setSiteManagerPermission(NodeRef nodeRef, NodePermissions nodePerms)
+    private String getSiteManagerAuthority(NodeRef nodeRef) {
+        return AuthenticationUtil.runAsSystem(() -> {
+            SiteInfo containingSite = siteService.getSite(nodeRef);
+
+            if (containingSite != null)
+            {
+                String thisSiteGroupPrefix = siteService.getSiteGroup(containingSite.getShortName());
+                return thisSiteGroupPrefix + "_" + SiteModel.SITE_MANAGER;
+            }
+
+            return null;
+        });
+    }
+
+    private void setSiteManagerPermission(NodeRef nodeRef, NodePermissions nodePerms, String siteManagerAuthority)
     {
-        if (nodeRef != null && nodePerms != null)
+        if (nodeRef != null && nodePerms != null && nodePerms.getIsInheritanceEnabled() != null && siteManagerAuthority != null)
         {
             try
             {
-                if (nodePerms.getIsInheritanceEnabled() != null && !nodePerms.getIsInheritanceEnabled())
+                if (!nodePerms.getIsInheritanceEnabled())
                 {
-                    SiteInfo containingSite = siteService.getSite(nodeRef);
-
-                    if (containingSite != null)
-                    {
-                        String thisSiteGroupPrefix = siteService.getSiteGroup(containingSite.getShortName());
-                        final String siteManagerAuthority = thisSiteGroupPrefix + "_" + SiteModel.SITE_MANAGER;
-                        AuthenticationUtil.runAsSystem(() -> {
-                            permissionService.setPermission(nodeRef, siteManagerAuthority, SiteModel.SITE_MANAGER, true);
-                            return null;
-                        });
-                    }
+                    permissionService.setPermission(nodeRef, siteManagerAuthority, SiteModel.SITE_MANAGER, true);
                 }
             }
             catch (Exception e)
